ci: Remove rustdoc job
The job was set up for building rustdocs, and pushing them to GitHub
pages. But since GitHub pages isn't enabled for this repository, it just
failed on our main branch each commit. I don't see much point in having
a hosted rustdoc for this repository, so we might as well delete it.